instance: generate valid uuid for machine-id
adding a prefix to the uuid that was being passed as a machine-id in the
fake client requests caused coreupdate to fail when it tried to parse
them as uuids. the solution is to pass it a valid uuid that still
fulfills the original objective of the prefix, namely to make fake
clients easy to pick out.

fixes #126