group: add group percent command
this subcommand sets the update percentage directly, instead of setting
it through the group update command. it also removes the update-percent
flag under the group update command so that it can no longer be used to
set the update percent. setting the update percent is semantically
different than setting other attributes of a group, so needs it's own
api call, since there is no good way to tell if that's what we are
setting in the group update api call. sorry!